// src/app/core/models/home-sections.ts
export interface LifestyleItem {
  image: string;
  alt: string;
  title?: string;
}

export interface SolutionCard {
  id: string;
  icon: string;
  title: string;
  description: string;
}
export interface SystemSlide {
  image: string;
  title: string;
  description: string;
  buttonText: string;
  buttonRoute: string;
}

export interface GridCards {
  title: string;
  description: string;
  image: string;
}
export interface Feature{
    text:string;
}

export interface Product {
  image: string;
  alt: string;
}